changes to style and redability
Changed Mono Font to IBM 3270, it looks nicer in this context and the styleguide is weird anyway. Changed the Sans Font to Red Hat Text, it reads a bit "heavier/thicker" which will probably work better in this context. Changed the colors, because it will be simpler to print black on white. Changed the Wording of a question to better fit the answer.